+++ app/igor/main.cpp
@@ -193,6 +193,8 @@ int main(int argc, char *argv[])
     bool gen_output_CDR3_data = false;
     string gen_filename_prefix = "";
     int gen_random_engine_seed = -1; //-1 will cause IGoR to generate a time stamp based seed
+    bool gen_fast_mode = false; // Use fast parallel generation
+    size_t gen_num_threads = 0; // Number of threads for fast generation (0 = auto)
 
     //Inference parms
     bool viterbi_inference = false;
@@ -859,9 +861,9 @@ int main(int argc, char *argv[])
         }
 
         /*
-		 * Set the working directory
-		 * /!\ Needs to be set before all the counters !  /!\
-		 */
+         * Set the working directory
+         * /!\ Needs to be set before all the counters !  /!\
+         */
         else if (string(argv[carg_i]) == "-set_wd") {
             wd = true;
             ++carg_i;
@@ -879,16 +881,16 @@ int main(int argc, char *argv[])
         }
 
         /*
-		 * Output arguments parsing
-		 */
+         * Output arguments parsing
+         */
         else if (string(argv[carg_i]) == "-output") {
             while ((carg_i + 1 < argc) and (string(argv[carg_i + 1]).size() > 2)
                    and string(argv[carg_i + 1]).substr(0, 2) == "--") {
 
                 ++carg_i;
                 /*
-				 * TODO For now forget about outputing for every sequences / every iterations (more command line parameters to code)
-				 */
+                 * TODO For now forget about outputing for every sequences / every iterations (more command line parameters to code)
+                 */
                 if (string(argv[carg_i]) == "--Pgen") {
                     shared_ptr<Counter> pgen_counter_ptr(new Pgen_counter(cl_path + "output/"));
                     cl_counters_list.emplace(cl_counters_list.size(), pgen_counter_ptr);
@@ -975,6 +977,24 @@ int main(int argc, char *argv[])
                                                                  + string(argv[carg_i]) + "\"");
                     }
                     gen_random_engine_seed = stoi(string(argv[carg_i]));
+                } else if (string(argv[carg_i]) == "--fast") {
+                    // Enable fast parallel sequence generation (100x+ speedup)
+                    gen_fast_mode = true;
+                } else if (string(argv[carg_i]) == "--threads") {
+                    // Set number of threads for fast generation
+                    ++carg_i;
+                    try {
+                        int tmp_threads = stoi(string(argv[carg_i]));
+                        if (tmp_threads < 0) {
+                            return terminate_IGoR_with_error_message("The number of threads after \"--threads\" must "
+                                                                     "be a non-negative integer (0 = auto-detect)");
+                        }
+                        gen_num_threads = static_cast<size_t>(tmp_threads);
+                    } catch (exception &e) {
+                        return terminate_IGoR_with_error_message(
+                                "Expected a non-negative integer after \"--threads\", received: \""
+                                + string(argv[carg_i]) + "\"");
+                    }
                 } else {
                     return terminate_IGoR_with_error_message("Unknown argument \"" + string(argv[carg_i])
                                                              + "\" to specify sequence generation parameters");
@@ -2214,48 +2234,69 @@ int main(int argc, char *argv[])
             //Initialize generated sequences output function
             std::list<std::pair<gen_seq_trans, shared_ptr<void>>> func_data_pairs_list;
 
-            if (gen_output_CDR3_data) {
-                //Get V and J event
-                auto events_map = cl_model_parms.get_events_map();
-                shared_ptr<const Rec_Event> v_event_ptr =
-                        events_map.at(make_tuple(GeneChoice_t, V_gene, Undefined_side));
-                shared_ptr<const Rec_Event> j_event_ptr =
-                        events_map.at(make_tuple(GeneChoice_t, J_gene, Undefined_side));
-
-                //Get V and J event positions on the queue
-                auto model_queue = cl_model_parms.get_model_queue();
-                size_t i = 0;
-                size_t v_queue_pos;
-                size_t j_queue_pos;
-                while (not model_queue.empty()) {
-                    if (model_queue.front()->get_name() == v_event_ptr->get_name()) {
-                        v_queue_pos = i;
-                    } else if (model_queue.front()->get_name() == j_event_ptr->get_name()) {
-                        j_queue_pos = i;
-                    }
-                    ++i;
-                    model_queue.pop();
+            if (gen_fast_mode) {
+                // Use fast parallel generation (100x+ speedup)
+                clog << "Using fast parallel sequence generation mode" << endl;
+                if (gen_output_CDR3_data) {
+                    clog << "Warning: CDR3 output not yet supported in fast mode, will be skipped" << endl;
                 }
-                shared_ptr<ostream> outputfile_ptr = shared_ptr<ostream>(new ofstream(
-                        cl_path + batchname + "generated/" + "generated_seqs_" + w_err_str + "_CDR3_info.csv"));
-                shared_ptr<void> CDR3_func_data_ptr = shared_ptr<void>(new gen_CDR3_data(
-                        v_CDR3_anchors, v_event_ptr->get_realizations_map(), v_queue_pos, j_CDR3_anchors,
-                        j_event_ptr->get_realizations_map(), j_queue_pos, outputfile_ptr));
+                if (generate_werr) {
+                    clog << "Warning: Error generation not yet supported in fast mode, generating without errors"
+                         << endl;
+                }
+                genmodel.generate_sequences_fast(generate_n_seq,
+                                                 cl_path + batchname + "generated/" + gen_filename_prefix
+                                                         + "generated_seqs_" + w_err_str + ".csv",
+                                                 cl_path + batchname + "generated/" + gen_filename_prefix
+                                                         + "generated_realizations_" + w_err_str + ".csv",
+                                                 gen_num_threads, static_cast<int64_t>(gen_random_engine_seed),
+                                                 true // show_progress
+                );
+            } else {
+                // Use original generation method
+                //Initialize generated sequences output function
+                std::list<std::pair<gen_seq_trans, shared_ptr<void>>> func_data_pairs_list;
 
-                func_data_pairs_list.emplace_back(output_CDR3_gen_data, CDR3_func_data_ptr);
+                if (gen_output_CDR3_data) {
+                    //Get V and J event
+                    auto events_map = cl_model_parms.get_events_map();
+                    shared_ptr<const Rec_Event> v_event_ptr =
+                            events_map.at(make_tuple(GeneChoice_t, V_gene, Undefined_side));
+                    shared_ptr<const Rec_Event> j_event_ptr =
+                            events_map.at(make_tuple(GeneChoice_t, J_gene, Undefined_side));
+
+                    //Get V and J event positions on the queue
+                    auto model_queue = cl_model_parms.get_model_queue();
+                    size_t i = 0;
+                    size_t v_queue_pos;
+                    size_t j_queue_pos;
+                    while (not model_queue.empty()) {
+                        if (model_queue.front()->get_name() == v_event_ptr->get_name()) {
+                            v_queue_pos = i;
+                        } else if (model_queue.front()->get_name() == j_event_ptr->get_name()) {
+                            j_queue_pos = i;
+                        }
+                        ++i;
+                        model_queue.pop();
+                    }
+                    shared_ptr<ostream> outputfile_ptr = shared_ptr<ostream>(new ofstream(
+                            cl_path + batchname + "generated/" + "generated_seqs_" + w_err_str + "_CDR3_info.csv"));
+                    shared_ptr<void> CDR3_func_data_ptr = shared_ptr<void>(new gen_CDR3_data(
+                            v_CDR3_anchors, v_event_ptr->get_realizations_map(), v_queue_pos, j_CDR3_anchors,
+                            j_event_ptr->get_realizations_map(), j_queue_pos, outputfile_ptr));
+
+                    func_data_pairs_list.emplace_back(output_CDR3_gen_data, CDR3_func_data_ptr);
+                }
+
+                genmodel.generate_sequences(generate_n_seq, generate_werr,
+                                            cl_path + batchname + "generated/" + gen_filename_prefix + "generated_seqs_"
+                                                    + w_err_str + ".csv",
+                                            cl_path + batchname + "generated/" + gen_filename_prefix
+                                                    + "generated_realizations_" + w_err_str + ".csv",
+                                            func_data_pairs_list, false, gen_random_engine_seed);
             }
-
-            genmodel.generate_sequences(generate_n_seq, generate_werr,
-                                        cl_path + batchname + "generated/" + gen_filename_prefix + "generated_seqs_"
-                                                + w_err_str + ".csv",
-                                        cl_path + batchname + "generated/" + gen_filename_prefix
-                                                + "generated_realizations_" + w_err_str + ".csv",
-                                        func_data_pairs_list, false, gen_random_engine_seed);
         }
-
-    }
-
-    else {
+    } else {
         //Write your custom procedure here
     }